Create separate component for Edit Box
