+++ client/src/views/Home.tsx
@@ -1,25 +1,18 @@
 import React, { useEffect, useState } from 'react';
 
 // Libraries
-import {
-  Button,
-  Box,
-  makeStyles,
-  FormControlLabel,
-  Switch,
-  Select,
-  MenuItem,
-} from '@material-ui/core';
-import ImageUploader from 'react-images-upload';
-import firebase from 'firebase/app';
-import 'firebase/firestore';
+import { Box, makeStyles } from '@material-ui/core';
 import jimp from 'jimp';
 import download from 'downloadjs';
 import html2canvas from 'html2canvas';
 
+// Firebase
+import firebase from 'firebase/app';
+import 'firebase/firestore';
+
 // Components
 import FrameCarousel from '../components/carousel/Carousel';
-import CustomTextField from '../components/shared/TextField';
+import EditBox from '../components/homepage/EditBox';
 
 // Utils
 import getCroppedImage from '../utils/cropImage';
@@ -28,24 +21,15 @@ import { determineTextboxDimensions } from '../utils/helpers';
 // Assets
 import { FrameData } from '../utils/types';
 
-// @ts-ignore
-function readFile(file) {
-  return new Promise((resolve) => {
-    const reader = new FileReader();
-    reader.addEventListener('load', () => resolve(reader.result), false);
-    reader.readAsDataURL(file);
-  });
-}
-
 const App: React.FC = () => {
-  const [uploadImage, setUploadImage] = useState<null | File>(null);
+  const [uploadImage, setUploadImage] = useState<string>('');
   const [crop, setCrop] = useState<{ x: number; y: number }>({ x: 0, y: 0 });
   const [zoom, setZoom] = useState<number>(1);
   const [aspect] = useState<number>(1 / 1);
 
   const [croppedAreaPixels, setCroppedAreaPixels] = useState<any>(null);
   const [frame, setFrame] = useState<FrameData | null>(null);
-  const [greyscale, setGreyScale] = useState<boolean>(false);
+  const [greyscale, setGreyscale] = useState<boolean>(false);
   const [textBoxDimensions, setTextBoxDimensions] = useState<any>();
 
   const [primaryText, setPrimaryText] = useState<string>('Primary Text');
@@ -71,26 +55,6 @@ const App: React.FC = () => {
         setFrame(frames[0]);
       });
   }, []);
-
-  // Event Handlers
-  const handlePositionChange = (event: React.ChangeEvent<{ value: unknown }>) =>
-    setPosition(event.target.value as string);
-
-  const handleGreyscaleChange = (
-    event: React.ChangeEvent<HTMLInputElement>
-  ) => {
-    setGreyScale(event.target.checked);
-  };
-
-  const handleImageUpload = async (e: any) => {
-    // const selected = e.target.files[0];
-    const selected = e[0];
-    if (selected) {
-      let imageDataUrl = await readFile(selected);
-      // @ts-ignore
-      setUploadImage(imageDataUrl);
-    }
-  };
 
   // Generate Frame
   const overlayImage = async () => {
@@ -162,78 +126,18 @@ const App: React.FC = () => {
 
   return (
     <div style={{ minHeight: window.innerHeight }}>
-      <Box className={classes.features}>
-        <FormControlLabel
-          control={
-            <Switch
-              checked={greyscale}
-              onChange={handleGreyscaleChange}
-              name='greyscale'
-              color='primary'
-            />
-          }
-          label='Grayscale'
-        />
-
-        <CustomTextField
-          value={primaryText}
-          setValue={setPrimaryText}
-          label='Primary Text'
-          type='text'
-          className={classes.textInput}
-        />
-
-        <CustomTextField
-          value={secondaryText}
-          setValue={setSecondaryText}
-          label='Secondary Text'
-          type='text'
-          className={classes.textInput}
-        />
-
-        <Select
-          label='Frame Shape'
-          onChange={handlePositionChange}
-          value={position}
-          variant='outlined'
-          style={{ width: '100%', marginTop: 20 }}
-        >
-          <MenuItem value='top-right'>Top Right</MenuItem>
-          <MenuItem value='top-left'>Top Left</MenuItem>
-          <MenuItem value='bottom-right'>Bottom Right</MenuItem>
-          <MenuItem value='bottom-left'>Bottom Left</MenuItem>
-        </Select>
-
-        <ImageUploader
-          className={classes.upload}
-          withIcon={false}
-          withLabel={false}
-          buttonText='CHOOSE IMAGE'
-          imgExtension={['.jpg', '.gif', '.png', '.gif', '.jpeg']}
-          maxFileSize={5242880 * 2}
-          singleImage={true}
-          fileContainerStyle={{
-            height: 'fit-content',
-            boxShadow: 'none',
-            backgroundColor: 'transparent',
-          }}
-          buttonClassName={classes.buttonStyles}
-          onChange={handleImageUpload}
-        />
-
-        <input
-          id='input'
-          type='file'
-          accept='image/png image/jpeg image/jpg'
-          alt='Image'
-          placeholder='Upload Image'
-          onChange={handleImageUpload}
-        />
-
-        <Button onClick={overlayImage} variant='contained' color='secondary'>
-          Download
-        </Button>
-      </Box>
+      <EditBox
+        primaryText={primaryText}
+        secondaryText={secondaryText}
+        position={position}
+        greyscale={greyscale}
+        setPrimaryText={setPrimaryText}
+        setSecondaryText={setSecondaryText}
+        setPosition={setPosition}
+        setGreyscale={setGreyscale}
+        setUploadImage={setUploadImage}
+        overlayImage={overlayImage}
+      />
 
       <Box className={classes.frame}>
         {data !== null ? (
@@ -263,44 +167,9 @@ const App: React.FC = () => {
 export default App;
 
 const useStyles = makeStyles((theme) => ({
-  features: {
-    transform: 'translate(-50%,-50%)',
-    position: 'absolute',
-    right: '5%',
-    top: '50%',
-    zIndex: 10,
-    height: '60%',
-    width: '350px',
-    backgroundImage: '#334d50',
-    background: 'linear-gradient(to right, #FDC830, #F37335)',
-    padding: '1rem',
-    borderRadius: '5px',
-    color: '#fff',
-    display: 'flex',
-    flexDirection: 'column',
-    marginBottom: 20,
-    [theme.breakpoints.down('sm')]: {
-      left: '50%',
-      top: '70%',
-      width: '320px',
-    },
-  },
   frame: {
     width: '100%',
     minHeight: window.innerHeight + 50,
     zIndex: 0,
-  },
-  upload: {
-    width: 'auto',
-  },
-  buttonStyles: {
-    backgroundColor: '#3F51B5',
-    width: '100%',
-    height: 'fit-content',
-    margin: 0,
-  },
-  textInput: {
-    marginTop: 20,
-    color: '#fff',
   },
 }));
